# openapi_spec.yml
openapi: 3.1.0
info:
  title: Advanced Real-Time Discussion Forum API
  description: Backend API for the discussion forum supporting JWT auth, RBAC, threads,
    nested comments, likes, notifications, moderation, and websocket updates.
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: Local backend
- url: http://localhost
  description: Nginx reverse proxy
paths:
  /api/v1/auth/register:
    post:
      tags:
      - Auth
      summary: Register User
      description: Register new user account.
      operationId: register_user_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - Auth
      summary: Login User
      description: Authenticate user and issue JWT tokens.
      operationId: login_user_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/refresh:
    post:
      tags:
      - Auth
      summary: Refresh Token
      operationId: refresh_token_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/change-password:
    post:
      tags:
      - Auth
      summary: Change Password
      operationId: change_password_api_v1_auth_change_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - Auth
      summary: Get Current User Profile
      description: Get authenticated user profile.
      operationId: get_current_user_profile_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - HTTPBearer: []
    put:
      tags:
      - Auth
      summary: Update Current User Profile
      operationId: update_current_user_profile_api_v1_auth_me_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/comments:
    post:
      tags:
      - Comments
      summary: Create Comment
      operationId: create_comment_api_v1_comments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/comments/thread/{thread_id}:
    get:
      tags:
      - Comments
      summary: List Comments
      operationId: list_comments_api_v1_comments_thread__thread_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: integer
          title: Thread Id
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentResponse'
                title: Response List Comments Api V1 Comments Thread  Thread Id  Get
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/comments/{comment_id}:
    put:
      tags:
      - Comments
      summary: Update Comment
      operationId: update_comment_api_v1_comments__comment_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: comment_id
        in: path
        required: true
        schema:
          type: integer
          title: Comment Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Comments
      summary: Delete Comment
      operationId: delete_comment_api_v1_comments__comment_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: comment_id
        in: path
        required: true
        schema:
          type: integer
          title: Comment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/likes:
    post:
      tags:
      - Likes
      summary: Add Like
      operationId: add_like_api_v1_likes_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LikeResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
    delete:
      tags:
      - Likes
      summary: Remove Like
      operationId: remove_like_api_v1_likes_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/search/comments:
    get:
      tags:
      - Search
      summary: Search Comments
      operationId: search_comments_api_v1_search_comments_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          title: Q
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentSearchResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AccessTokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
      type: object
      required:
      - access_token
      title: AccessTokenResponse
...
    ChangePasswordRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        old_password:
          type: string
          title: Old Password
        new_password:
          type: string
          title: New Password
      type: object
      required:
      - email
      - old_password
      - new_password
      title: ChangePasswordRequest
    CommentCreate:
      properties:
        thread_id:
          type: integer
          title: Thread Id
        content:
          type: string
          title: Content
        parent_comment_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Parent Comment Id
      type: object
      required:
      - thread_id
      - content
      title: CommentCreate
    CommentResponse:
      properties:
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        content:
          type: string
          title: Content
        thread_id:
          type: integer
          title: Thread Id
        author_id:
          type: integer
          title: Author Id
        author:
          anyOf:
          - $ref: '#/components/schemas/ThreadAuthor'
          - type: 'null'
        parent_comment_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Parent Comment Id
        like_count:
          type: integer
          title: Like Count
          default: 0
        user_has_liked:
          type: boolean
          title: User Has Liked
          default: false
        is_deleted:
          type: boolean
          title: Is Deleted
      type: object
      required:
      - id
      - created_at
      - updated_at
      - content
      - thread_id
      - author_id
      - parent_comment_id
      - is_deleted
      title: CommentResponse
    CommentSearchResponse:
      properties:
        results:
          items:
            $ref: '#/components/schemas/CommentResponse'
          type: array
          title: Results
        total:
          type: integer
          title: Total
      type: object
      required:
      - results
      - total
      title: CommentSearchResponse
    CommentUpdate:
      properties:
        content:
          type: string
          title: Content
      type: object
      required:
      - content
      title: CommentUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LikeCreate:
      properties:
        thread_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Thread Id
        comment_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Comment Id
      type: object
      title: LikeCreate
    LikeResponse:
      properties:
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        user_id:
          type: integer
          title: User Id
        thread_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Thread Id
        comment_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Comment Id
      type: object
      required:
      - id
      - created_at
      - updated_at
      - user_id
      - thread_id
      - comment_id
      title: LikeResponse
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
      description: User login payload.
...
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: MessageResponse
...
    RefreshTokenRequest:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: RefreshTokenRequest
      description: Refresh token payload.
...
    RoleResponse:
      properties:
        id:
          type: integer
          title: Id
        role_name:
          type: string
          title: Role Name
      type: object
      required:
      - id
      - role_name
      title: RoleResponse
    ThreadAuthor:
      properties:
        id:
          type: integer
          title: Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        email:
          type: string
          title: Email
        avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar Url
      type: object
      required:
      - id
      - email
      title: ThreadAuthor
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
      type: object
      required:
      - access_token
      - refresh_token
      title: TokenResponse
      description: JWT token response.
...
    UserCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      type: object
      required:
      - email
      - password
      title: UserCreate
      description: Request schema for user registration.
...
    UserResponse:
      properties:
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        email:
          type: string
          title: Email
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar Url
        bio:
          anyOf:
          - type: string
          - type: 'null'
          title: Bio
        is_active:
          type: boolean
          title: Is Active
        roles:
          items:
            $ref: '#/components/schemas/RoleResponse'
          type: array
          title: Roles
          default: []
      type: object
      required:
      - id
      - created_at
      - updated_at
      - email
      - name
      - avatar_url
      - bio
      - is_active
      title: UserResponse
      description: Public user response schema.
...
    UserUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar Url
        bio:
          anyOf:
          - type: string
          - type: 'null'
          title: Bio
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
      type: object
      title: UserUpdate
      description: Update user profile.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...